requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1
django-json-widget==1.1.1
django-celery-beat==2.4.0
django-cors-headers==3.14.0
django-filter==23.1
django-import-export==3.1.0
django-jazzmin==2.6.0
django-timezone-field==5.0
djangorestframework==3.14.0
drf-yasg==1.21.5
et-xmlfile==1.1.0
flower==1.2.0
gunicorn==20.1.0
humanize==4.6.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
MarkupPy==1.14
MarkupSafe==2.1.2
mssql-django==1.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.1
packaging==23.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.1.0
pyodbc==4.0.35
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
tomli==2.0.1
tornado==6.2
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
watchdog==2.3.1
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
whitenoise==6.5.0